rename project phases` services and contracts
