+++ app/contracts/project_phases/base_contract.rb
@@ -1,5 +1,3 @@
-# frozen_string_literal: true
-
 #-- copyright
 # OpenProject is an open source project management software.
 # Copyright (C) the OpenProject GmbH
@@ -28,31 +26,14 @@
 # See COPYRIGHT and LICENSE files for more details.
 #++
 
-module ProjectLifeCycleSteps
-  class UpdateService < ::BaseServices::Update
-    delegate :project, to: :model
+module ProjectPhases
+  class BaseContract < ::ModelContract
+    validate :validate_edit_project_phases_permission
 
-    def after_perform(*)
-      reschedule_following_phases if model.date_range_set?
+    def validate_edit_project_phases_permission
+      return if user.allowed_in_project?(:edit_project_phases, project)
 
-      project.touch_and_save_journals
-
-      super
-    end
-
-    private
-
-    def reschedule_following_phases
-      RescheduleService.new(user:, project:)
-        .call(phases: following_phases, from: initial_reschedule_date)
-    end
-
-    def initial_reschedule_date
-      model.active? ? model.finish_date + 1 : model.start_date
-    end
-
-    def following_phases
-      project.available_phases.select { it.position > model.position }
+      errors.add :base, :error_unauthorized
     end
   end
 end
